feat(components): [date-picker] extract props (#7893)
- Extract props from basic-month-table.
- Extract common props to shared.
