+++ packages/components/date-picker/src/props/basic-month-table.ts
@@ -1,0 +1,11 @@
+import { buildProps } from '@element-plus/utils'
+import { datePickerSharedProps, selectionModeWithDefault } from './shared'
+
+import type { ExtractPropTypes } from 'vue'
+
+export const basicMonthTableProps = buildProps({
+  ...datePickerSharedProps,
+  selectionMode: selectionModeWithDefault('month'),
+})
+
+export type BasicMonthTableProps = ExtractPropTypes<typeof basicMonthTableProps>
